requirements.txt
python-dotenv>=1.0
pydantic-settings>=1.0
pydantic>=1.10
loguru>=0.7
openai>=2.17.0